rename generation route, clean up, hide Admin/demo features
