+++ app/generations/page.tsx
@@ -1,4 +1,4 @@
-import { Generations } from '@/app/generate-demo/Generations'
+import { Generations } from '@/app/generations/Generations'
 import { GenerationSection } from '@/components/generation/GenerationSection'
 
 export default function Page() {
