+++ lib/api.ts
@@ -250,6 +250,10 @@ export const useModels = (resourceKey?: string) => {
   return result as Partial<typeof result>
 }
 
+export const useViewer = () => {
+  return useQuery(api.users.getViewer, {})
+}
+
 // * collections
 export const useCollection = (collectionId?: Id<'collections'>) => {
   const collection = useQuery(api.db.collections.get, collectionId ? { collectionId } : 'skip')
